Don't look at the depth when determining wether an item is a current ancestor or not.
Saying that if the current child is not shown, then the parent is not an
ancestor, does not make sense.
The child can be shown in another menu call.
For exemple, if I render the menu with depth one, then the seconde level
separatly. I want the ancestor item on level one to be considered as
ancestor.