Adding a new "alias" provider, which allows you to easily create menus by following a convention
The new synta is AcmeMyBundle:Builder:mainMenu, which would instantiate a new instance of a Acme\MyBundle\Menu\Builder class (with no arguments), inject the container if it's container aware, and then call the mainMenu method on it, expecting to get back an ItemInterface. The menu factory is passed to the method so that the method has what it needs to create the menu.