Changing BuilderAliasProvider - each time you ask for a menu, it's new - not the same instance as before
This addresses an issue brought up by @Stof. Specifically, if the same menu were asked for in different scopes, the menu would e the same in both cases, but the second would be built with services from the scope of the former. This is an edge scenario, but I think this is more correct.