Implement a MenuBuilderProviderInterface to support autoconfiguration
This interface allows defining the menu names from the class itself instead
of having to define them in the service definition. This is similar to the
event subscriber vs event listener system in Symfony.
The benefit of this interface is that the service definition can rely on
autoconfiguration to register everything, which fits very well with the Flex
way.