Use a tagged iterator argument to configure the matcher
Letting Symfony process the collection of voters to build the
`IteratorArgument` in the definition gives us all the features available
in recent Symfony versions, while our own processing logic was not
implementing them. For instance, this gives us support for the
`[AsTaggedItem]` attribute as a way to specify the priority of the
voter, which is great when using autoconfiguration.